requirements.txt
google-generativeai>=0.3.0
chromadb>=0.4.0
numpy>=1.21.0
tqdm>=4.64.0
python-dotenv>=1.0.0 